update: clear counter pid after shutdown

+++ rspi/launch_pushup_counter.sh
@@ -12,10 +12,12 @@ source $current_dir/dest/env.sh
 
 # 1. kill previous counter process if exist
 if test -f "$pid_path"; then
-  echo "begin kill previous counter process#$prev_pid" >> $log_file
   prev_pid="`cat $pid_path`"
-  kill_msg=`kill -9 $prev_pid`
-  echo "done kill counter process: $kill_msg" >> $log_file
+  if [[ -n $prev_pid ]]; then
+    echo "begin kill previous counter process#$prev_pid" >> $log_file
+    kill_msg=`kill -9 $prev_pid`
+    echo "done kill counter process: $kill_msg" >> $log_file
+  fi
 fi
 
 # 2. run counter
@@ -34,3 +36,5 @@ fi
 echo "Push-up counter launched." >> $log_file
 
 echo "READY"
+
+exit 0
